refactor: rename the files in the utils
