+++ src/main-app/typings/globals.d.ts
@@ -1,5 +1,5 @@
 declare namespace NodeJS {
     export interface Global {
-        runtime: import("../src/utils/Runtime").Runtime;
+        runtime: import("../src/utils/runtime").Runtime;
     }
 }
