+++ src/renderer-app/src/AddNamePage.tsx
@@ -4,7 +4,7 @@ import { Link } from "react-router-dom";
 import "./AddNamePage.less";
 import logo from "./assets/image/logo.svg";
 import { Button, Input } from "antd";
-import { ipcAsyncByMain } from "./utils/Ipc";
+import { ipcAsyncByMain } from "./utils/ipc";
 import { Identity } from "./utils/localStorage/room";
 
 export type JoinPageStates = {
