+++ src/renderer-app/src/tasks/Init-ui.tsx
@@ -1,8 +1,8 @@
 import React from "react";
 import ReactDOM from "react-dom";
-import { AppRoutes } from "../AppRoutes";
 import "antd/dist/antd.less";
 import "../theme.less";
+import { AppRoutes } from "../AppRoutes";
 
 const initUI = () => {
     ReactDOM.render(<AppRoutes />, document.getElementById("root") as HTMLElement);
