+++ src/renderer-app/src/utils/i18n.ts
