+++ src/renderer-app/src/utils/wechatUrl.ts
